langgraph-single-agent-tools/requirements.txt
langgraph>=0.2.0
langchain-openai>=0.1.20
langchain-community>=0.2.0
langchain-experimental>=0.3.0
langchain-core>=0.2.0
python-dotenv>=1.0.0
ddgs>=9.0.0
numexpr>=2.8.0
